+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.0.0] - 2020-02-27
 
 ### Added
  - Province information to `Location`.
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `New` can now accept multiple datasets (which is only really useful when you
    want to use `Cities10` and get province/country information.
  - Switched to Apache License.
+ - Better errors.
+ - 100% test coverage.
 
 ## [0.0.5] - 2020-02-23